+++ pkg/common/event/sync_point_event_test.go
@@ -164,7 +164,7 @@ func TestSyncPointEventHeader(t *testing.T) {
 
 	// Verify total size
 	headerSize := GetEventHeaderSize()
-	require.Equal(t, headerSize+payloadLen, len(data))
+	require.Equal(t, uint64(headerSize)+payloadLen, uint64(len(data)))
 }
 
 // TestSyncPointEventUnmarshalErrors tests error handling in Unmarshal
